+++ src/components/Signup/index.tsx
@@ -1,0 +1,3 @@
+export { default as Email } from './Email'
+export { default as Password } from './Password'
+export { default as Nickname } from './Nickname'
