+++ src/pages/Signup/signup.module.scss
@@ -15,12 +15,8 @@
       }
     }
 
-    .duplicate {
+    .green {
       color: colors.$PRIMARY_TEXT;
-    }
-
-    .red {
-      color: colors.$RED;
     }
   }
