+++ src/types/index.d.ts
@@ -5,3 +5,4 @@ export * from './region'
 export * from './user'
 export * from './feed'
 export * from './apis'
+export * from './signup'
